+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemCrossbowStack.java
@@ -2,6 +2,7 @@ package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.item.component.ItemCrossbowBaseComponent;
+import org.allaymc.api.item.component.ItemRepairableComponent;
 
-public interface ItemCrossbowStack extends ItemStack, ItemCrossbowBaseComponent {
+public interface ItemCrossbowStack extends ItemStack, ItemCrossbowBaseComponent, ItemRepairableComponent {
 }