+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemMaceStack.java
@@ -1,7 +1,8 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemRepairableComponent;
 import org.allaymc.api.item.component.ItemToolComponent;
 
-public interface ItemMaceStack extends ItemStack, ItemToolComponent {
+public interface ItemMaceStack extends ItemStack, ItemToolComponent, ItemRepairableComponent {
 }