+++ server/src/main/java/org/allaymc/server/item/type/ItemTypeInitializer.java
@@ -979,4 +979,14 @@ public final class ItemTypeInitializer {
                 .addComponent(ItemComparatorBaseComponentImpl::new, ItemComparatorBaseComponentImpl.class)
                 .build();
     }
+
+    public static void initMace() {
+        ItemTypes.MACE = AllayItemType
+                .builder(ItemMaceStackImpl.class)
+                .vanillaItem(ItemId.MACE)
+                .addComponent(ItemMaceBaseComponentImpl::new, ItemMaceBaseComponentImpl.class)
+                .addComponent(ItemSwordToolComponentImpl::new, ItemSwordToolComponentImpl.class)
+                .addComponent(() -> new ItemRepairableComponentImpl(ItemId.BREEZE_ROD), ItemRepairableComponentImpl.class)
+                .build();
+    }
 }